requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
more-itertools==4.1.0
mypy==0.580
pluggy==0.6.0
py==1.5.3
pytest==3.5.0
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
typed-ast==1.1.0
urllib3==1.22
